refactor #6 - 디자인 코드를 별도의 ViewController로 옮김
